move test involving docker from `test_ipython` to `test_sandbox`
